using int32 gives a little performance

+++ src/segs.c
@@ -2,14 +2,14 @@
 #include <stdbool.h>
 #include <math.h>
 
-double min(double [4]);
-double point_seg_dist(double px, double py, double *line);
-bool segs_intersect(double *line0, double *line1);
+int32_t min(int32_t [4]);
+int32_t point_seg_dist(int32_t px, int32_t py, int32_t *line);
+bool segs_intersect(int32_t *line0, int32_t *line1);
 
-double segs(double * restrict line0, double * restrict line1){
+int32_t segs(int32_t * restrict line0, int32_t * restrict line1){
     if (segs_intersect(line0, line1))
         return 0;
-    double distances[4];
+    int32_t distances[4];
     distances[0] = point_seg_dist(line0[0], line0[1], line1);
     distances[1] = point_seg_dist(line0[2], line0[3], line1);
     distances[2] = point_seg_dist(line1[0], line1[1], line0);
@@ -17,7 +17,7 @@ double segs(double * restrict line0, double * restrict line1){
     return min(distances);
 }
 
-bool segs_intersect(double *line0, double *line1) {
+bool segs_intersect(int32_t *line0, int32_t *line1) {
     double x0, y0, x1, y1;
     double xx0, yy0, xx1, yy1;
     x0 = line0[0];
@@ -29,12 +29,12 @@ bool segs_intersect(double *line0, double *line1) {
     xx1 = line1[2];
     yy1 = line1[3];
     
-    double dx0 = x1 - x0;
-    double dy0 = y1 - y0;
-    double dx1 = xx1 - xx0;
-    double dy1 = yy1 - yy0;
+    int32_t dx0 = x1 - x0;
+    int32_t dy0 = y1 - y0;
+    int32_t dx1 = xx1 - xx0;
+    int32_t dy1 = yy1 - yy0;
 
-    int delta = dx1*dy0 - dy1*dx0;
+    int32_t delta = dx1*dy0 - dy1*dx0;
     if (delta == 0)
         return false;
 
@@ -46,7 +46,7 @@ bool segs_intersect(double *line0, double *line1) {
     return ss && tt;
 }
 
-double point_seg_dist(double px, double py, double *line) {
+int32_t point_seg_dist(int32_t px, int32_t py, int32_t *line) {
     double x0, y0, x1, y1;
     x0 = line[0];
     y0 = line[1];
@@ -77,7 +77,7 @@ double point_seg_dist(double px, double py, double *line) {
     return sqrt(dx*dx + dy*dy);
 }
 
-double min(double distances[4]) {
+int32_t min(int32_t distances[4]) {
     int i = 0;
     double m = distances[i];
     while (++i < 4) {
